refactor dtos and mappings
