+++ src/application/organization/test/organization.integration.spec.ts
@@ -4,13 +4,12 @@ import supertest from 'supertest';
 import { container } from 'tsyringe';
 
 import { Server } from "../../../infrastructure/config/server";
-import { OrganizationDto } from "../organizationDto";
+import { OrganizationDto } from "../dtos/organizationDto";
 import { DbConnection } from "../../../infrastructure/config/dbConnection";
 import { UsersRepository } from "../../users/usersRepository";
 import { CryptoService } from "../../../infrastructure/utils/CryptoService";
 import { generateUserModel } from "../../../infrastructure/utils/Models";
 import { auth } from '../../../infrastructure/config/authConfiguration';
-import { UserRole } from '../../../domain/entities/User';
 
 // Helper function to generate a mock organization model
 const generateOrganizationModel = () => {
