+++ src/application/security/test/security.unit.spec.ts
@@ -3,12 +3,9 @@ import { expect } from "chai";
 import { SecurityController } from "../securityController";
 import { SecurityService } from "../securityService";
 import { mock, instance, when, anyString, anything, verify } from "ts-mockito";
-import * as jwt from "jsonwebtoken";
-import { IloginDto } from "../Dtos/loginDto";
-import { ISignupDto } from "../Dtos/signupDto";
-import { ISecurityDto } from "../Dtos/securityDto";
-import constants from "../../../infrastructure/config/constants";
-import { generateUserModel } from "../../../infrastructure/utils/Models";
+import { IloginDto } from "../dtos/loginDto";
+import { ISignupDto } from "../dtos/signupDto";
+import { ISecurityDto } from "../dtos/securityDto";
 import { UserRole } from "../../../domain/entities/User";
 import { auth } from "../../../infrastructure/config/authConfiguration";
 import sinon from "sinon";
