+++ src/application/users/dtos/userDto.ts
@@ -1,4 +1,4 @@
-import { UserRole } from "../../domain/entities/User";
+import { UserRole } from "../../../domain/entities/User";
 import { AutoMap } from '@automapper/classes';
 
 export interface IUserDto {
