+++ src/application/users/test/users.unit.spec.ts
@@ -2,7 +2,7 @@ import "reflect-metadata";
 import { expect } from "chai";
 import { UsersService } from "../usersService";
 import { UsersRepository } from "../usersRepository";
-import { UserDto } from "../userDto";
+import { UserDto } from "../dtos/userDto";
 import { PaginationDto } from "../../../infrastructure/utils/PaginationDto";
 import { generateUserModel, generateMockUUID } from "../../../infrastructure/utils/Models";
 import { mock, instance, when, anyString, anything } from "ts-mockito";
